feat: add owner authorization to handler functions.

+++ src/api/notes/handler.js
@@ -16,8 +16,11 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { title = 'untitled', body, tags } = request.payload;
- 
-      const noteId = await this._service.addNote({ title, body, tags });
+      const { id: credentialId } = request.auth.credentials;
+
+      const noteId = await this._service.addNote({
+        title, body, tags, owner: credentialId,
+      });
  
       const response = h.response({
         status: 'success',
@@ -49,8 +52,9 @@ class NotesHandler {
     }
   }
  
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -62,6 +66,9 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+    
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
       return {
         status: 'success',
@@ -90,13 +97,15 @@ class NotesHandler {
     }
   }
  
-  putNoteByIdHandler(request, h) {
+  async putNoteByIdHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
       const { id } = request.params;
- 
-      this._service.editNoteById(id, request.payload);
- 
+      const { id: credentialId} = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
+      await this._service.editNoteById(id, request.payload);
+
       return {
         status: 'success',
         message: 'Catatan berhasil diperbarui',
@@ -110,7 +119,7 @@ class NotesHandler {
         response.code(error.statusCode);
         return response;
       }
- 
+
       // Server ERROR!
       const response = h.response({
         status: 'error',
@@ -122,11 +131,13 @@ class NotesHandler {
     }
   }
  
-  deleteNoteByIdHandler(request, h) {
+  async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
-      this._service.deleteNoteById(id);
- 
+      const { id: credentialId } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
+      await this._service.deleteNoteById(id);
+  
       return {
         status: 'success',
         message: 'Catatan berhasil dihapus',
@@ -140,7 +151,7 @@ class NotesHandler {
         response.code(error.statusCode);
         return response;
       }
- 
+  
       // Server ERROR!
       const response = h.response({
         status: 'error',
